[SPARK-7481]. spark-cloud is no longer a module built on hadoop-2.6; s3a performance and the wontfix issue of FS-impl-load-failures means that you can't use s3a in it, but that you wouldn't be able to init Hadoop's FileSystem class without the amazon-sdk on the classpath. It also guarantees that the hadoop-azure module exists.
Hadoop 2.7.2 has the same initialization problem, but the next 2.7.x release will fix that.

The assembly no-longer strips out the amazon and azure libraries, so that they end up in SPARK_HOME/jars and automatically uploaded. This elminates the problem of getting those JARs into place. It will make the JAR a lot bigger though...maybe it would be best to (initially) build and include the spark-cloud module if a -Pcloud profile were to be used.